# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: iSmart Gravure Management API
  description: Consolidated API contracts and endpoints for the Gravure Management System.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Next.js API Proxy Gateway (Production/Dev Default)
  - url: http://localhost:5000/api/v1
    description: Direct Backend Monolith Service
paths:
...
  /customers:
    get:
      summary: Get List of Customers
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval
    post:
      summary: Create New Customer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name
              properties:
                code:
                  type: string
                  example: CUST-EXCEL
                name:
                  type: string
                  example: บริษัท เอ็กซ์เซล ฟู้ดส์
                contactInfo:
                  type: string
      responses:
        '201':
          description: Created successfully
...
  /products:
    get:
      summary: Get List of Products
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval
    post:
      summary: Create New Product
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name
                - customerCode
              properties:
                code:
                  type: string
                  example: AGH-001
                name:
                  type: string
                  example: AGH-001 Package
                customerCode:
                  type: string
                  example: CUST-EXCEL
      responses:
        '201':
          description: Created successfully
...
  /cylinders:
    get:
      summary: Get List of Cylinders
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval
    post:
      summary: Create New Cylinder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - productCode
                - color
                - colorName
                - location
                - size
              properties:
                id:
                  type: string
                  example: CYL-BK-001
                productCode:
                  type: string
                  example: AGH-001
                color:
                  type: string
                  example: BK
                colorName:
                  type: string
                  example: Black
                location:
                  type: string
                  example: Rack A-03
                type:
                  type: string
                  example: Dedicated
                size:
                  type: string
                  example: 800×250
      responses:
        '201':
          description: Created successfully
...
  /orders:
    get:
      summary: Get List of Sales Orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful retrieval of sales orders
    post:
      summary: Create New Sales Order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderNumber
                - customerCode
                - productCode
                - quantity
                - unit
                - dueDate
              properties:
                orderNumber:
                  type: string
                  example: SO2024-001
                customerCode:
                  type: string
                  example: CUST-EXCEL
                productCode:
                  type: string
                  example: AGH-001
                quantity:
                  type: number
                  example: 25000
                unit:
                  type: string
                  example: m
                dueDate:
                  type: string
                  format: date-time
                  example: "2024-06-25T00:00:00.000Z"
      responses:
        '201':
          description: Created successfully
...
  /jobs:
    get:
      summary: Get List of Production Jobs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful retrieval
    post:
      summary: Create New Production Job
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobNumber
                - orderNumber
                - productCode
                - machineName
                - plannedDate
              properties:
                jobNumber:
                  type: string
                  example: J2024-045
                orderNumber:
                  type: string
                  example: SO2024-001
                productCode:
                  type: string
                  example: AGH-001
                machineName:
                  type: string
                  example: M-03
                plannedDate:
                  type: string
                  format: date-time
                  example: "2024-06-20T00:00:00.000Z"
      responses:
        '201':
          description: Created successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...